Adapt FWTS build for 32b arm
We add a patch to simplify what we build in FWTS for 32b arm, and to enable
UEFI support on 32b arm.
We supply a copy of libbsd's strlcpy() and strlcat() functions.
We copy the libraries fwts needs at runtime from the compiler.
This repairs the build for 32b arm and is sufficient to test for EBBR.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>